+++ src/js/views/main/main.js
@@ -1,14 +1,19 @@
-var Ractive = require('ractive')
-  , lMap = require('../../components/map/map')
+var Ractive   = require('ractive')
+  , lMap      = require('components/map/map')
+  , pieChart  = require('components/charts/PieChart')
   , main
 ;
 
 main = Ractive.extend({
   el: '#js-app',
   template: require('./template.html'),//temp,//'<div>{{message}}</div>',
+  data: {
+    initialData: 'world_innvandring.json'
+  },
 
   components: {
-    map: lMap
+    Map: lMap,
+    PieChart: pieChart
   },
 
   oninit: function () {