+++ CHANGELOG.md
@@ -5,15 +5,17 @@ This project adheres (well, maybe not quite) to [Semantic Versioning](http://sem
 
 
 
-## [UNRELEASED]
+## [0.0.1] - 2016-06-01
 ### Added
 - Initial package.json and build scripts
 - Added leaflet map component
 - Added copy leaflet.css on build
+- Add d3by5-pie-chart to library
+- Add Component that creates Pie Charts
 
 
 [//]: ##############################################
-
+<!---
 [//]: # (Legend)
 [Added]:        <> (for new features.)
 [Changed]:      <> (for changes in existing functionality.)
@@ -21,4 +23,4 @@ This project adheres (well, maybe not quite) to [Semantic Versioning](http://sem
 [Removed]:      <> (for deprecated features removed in this release.)
 [Fixed]:        <> (for any bug fixes.)
 [Security]:     <> (to invite users to upgrade in case of vulnerabilities.)
-
+--->